MyFile.java datatype issue is solved

+++ Server/src/MyFile.java
@@ -1,18 +1,18 @@
 public class MyFile {
 
-    private int id;
+    private long id;
     private  String name;
     private  byte[] data;
     private String fileExtension;
 
-    public MyFile(int id, String name, byte[] data, String fileExtension) {
+    public MyFile(long id, String name, byte[] data, String fileExtension) {
         this.id = id;
         this.name = name;
         this.data = data;
         this.fileExtension = fileExtension;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
